Fix https://github.com/synfig/synfig/issues/909
